plugins/hiagent-plugin-volc-speech2text/pyproject.toml
[tool.poetry]
authors = []
name = "hiagent-plugin-volc-speech2text"
version = "0.2.1"
description = ""
package-mode = true
include = [
  "hiagent_plugin_volc_speech2text/icon.png",
  "hiagent_plugin_volc_speech2text/*_metadata.yaml",
]

[tool.poetry.scripts]
gen_metadata = 'hiagent_plugin_volc_speech2text:VolcSpeech2Text._gen_metadata'

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "^24.1.0"
aiohttp = "^3.11.2"
websockets = "^15.0.1"

[tool.poetry.group.dev.dependencies]
hiagent_plugin_sdk = { path = "../../hiagent-plugin-sdk" }
python-dotenv = "^1.0.1"
types-aiofiles = "^24.1.0.20240626"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."hiagent.plugins"]
"录音文件识别" = "hiagent_plugin_volc_speech2text:VolcSpeech2Text"
